Add new endpoints, update SeedRoles, and modify DB schema

+++ Services/Wallet/IWalletTransactionService.cs
@@ -9,6 +9,6 @@ namespace Services.Wallet
 {
     public interface IWalletTransactionService
     {
-        Task AddWalletTransactionAsync(Guid walletId, int amount, string transactionType, string transactionStatus, string transactionDate, int transacionBalance, int? orderId);
+        Task AddWalletTransactionAsync(Guid walletId, decimal amount, string transactionType, string transactionStatus, string transactionDate, decimal transacionBalance, int? orderId);
     }
 }
